expose the AbstractTicketException in the login webflow
This will allow custom webflow logic to trigger for certain TicketExceptions

+++ cas-server-webapp-actions/src/main/java/org/jasig/cas/web/flow/AuthenticationViaFormAction.java
@@ -3,11 +3,11 @@ package org.jasig.cas.web.flow;
 import org.apache.commons.lang3.StringUtils;
 import org.jasig.cas.CasProtocolConstants;
 import org.jasig.cas.CentralAuthenticationService;
+import org.jasig.cas.authentication.AuthenticationException;
 import org.jasig.cas.authentication.AuthenticationResult;
 import org.jasig.cas.authentication.AuthenticationSystemSupport;
-import org.jasig.cas.authentication.DefaultAuthenticationSystemSupport;
-import org.jasig.cas.authentication.AuthenticationException;
 import org.jasig.cas.authentication.Credential;
+import org.jasig.cas.authentication.DefaultAuthenticationSystemSupport;
 import org.jasig.cas.authentication.HandlerResult;
 import org.jasig.cas.authentication.MessageDescriptor;
 import org.jasig.cas.authentication.principal.Service;
@@ -48,6 +48,8 @@ public class AuthenticationViaFormAction {
     /** Authentication failure result. */
     public static final String AUTHENTICATION_FAILURE = "authenticationFailure";
 
+    /** Ticket failure result. */
+    private static final String TICKET_FAILURE = "ticketFailure";
 
     /** Flow scope attribute that determines if authn is happening at a public workstation. */
     public static final String PUBLIC_WORKSTATION_ATTRIBUTE = "publicWorkstation";
@@ -165,14 +167,13 @@ public class AuthenticationViaFormAction {
 
         } catch (final AuthenticationException e) {
             return newEvent(AUTHENTICATION_FAILURE, e);
-        } catch (final TicketCreationException e) {
-            logger.warn("Invalid attempt to access service using renew=true with different credential. Ending SSO session.");
-            this.centralAuthenticationService.destroyTicketGrantingTicket(ticketGrantingTicketId);
         } catch (final AbstractTicketException e) {
-            return newEvent(AbstractCasWebflowConfigurer.TRANSITION_ID_ERROR, e);
+            if (e instanceof TicketCreationException) {
+                logger.warn("Invalid attempt to access service using renew=true with different credential. Ending SSO session.");
+                this.centralAuthenticationService.destroyTicketGrantingTicket(ticketGrantingTicketId);
+            }
+            return newEvent(TICKET_FAILURE, e);
         }
-        return newEvent(AbstractCasWebflowConfigurer.TRANSITION_ID_ERROR);
-
     }
     /**
      * Create ticket granting ticket for the given credentials.
